style: rename the mocks folder to sample-data
