+++ src/application/use-cases/calculate-eligibility-score.ts
@@ -1,11 +1,11 @@
 import { IProfessional } from '@/domain/entities/utils/professional.interface';
 import { Professional } from '@/domain/entities/professional';
 import { IEducationLevel } from '@/domain/entities/utils/educational-level.interface';
-import * as educationLevelsData from '@/infrastructure/mocks/education-levels.mock.json';
+import * as educationLevelsData from '@/infrastructure/sample-data/education-levels.mock.json';
 import { PastExperiences } from '@/domain/entities/utils/past-experiences';
 import { InternetTest } from '@/domain/entities/utils/internet-test';
 import { IReferralCode } from '@/domain/entities/utils/referral-code.interface';
-import * as referralCodeData from '@/infrastructure/mocks/referral-codes.mock.json';
+import * as referralCodeData from '@/infrastructure/sample-data/referral-codes.mock.json';
 
 export class CalculateEligibilityScore {
   public async execute(data: IProfessional): Promise<number> {
